Fix issue encoding hash character when used in a file name (#175)
Using `URI(protocol, null, host, port, path, query, ref)` encodes the path correctly except for the `#` character which is confusing. The fragment is sent in the `ref` argument, so it should know to escape it. 

Switch to the encoder provided in `com.google.common.net` and encode the substitutions manually.